Stop using Sass glob importing
Glob importing is a feature of sass-rails. It makes for cleanly and
easily importing lots of partials, but it completely hides Sass errors.

For example, when you miss a semicolon and your Sass doesn't compile,
using glob imports with sass-rails you simply see an application error:
"We're sorry, but something went wrong."

But when using standard Sass imports, this error is an actual syntax
error, which provides helpful messaging ("Invalid CSS after…"), and
traces back to the source where the error was introduced.